### 2023-10-15 #5951 @editorialbot
[PRE REVIEW]: Jobflow: Computational Workflows Made Simple
**Submitting author:** @Andrew-S-Rosen (<a href="http://orcid.org/0000-0002-0141-7006">Andrew S. Rosen</a>)
**Repository:** https://github.com/materialsproject/jobflow
**Branch with paper.md** (empty if default branch): main
**Version:** 0.1.14
