+++ paper/refs.bib
@@ -16,7 +16,8 @@ url = {https://github.com/Quantum-Accelerators/quacc}
   volume={1},
   number={1},
   year={2013},
-  publisher={AIP Publishing}
+  publisher={AIP Publishing},
+  doi={10.1063/1.4812323},
 }
 @software{rxnnetwork,
 title = {Reaction Network},
@@ -30,7 +31,8 @@ journal={Computational Materials Science},
 volume={111},
 pages={218--230},
 year={2016},
-publisher={Elsevier}
+publisher={Elsevier},
+doi={10.1016/j.commatsci.2015.09.013},
 }
 @article{mcdermott2021graph,
 title={A graph-based network for predicting chemical reaction pathways in solid-state materials synthesis},
@@ -40,7 +42,8 @@ volume={12},
 number={1},
 pages={3097},
 year={2021},
-publisher={Nature Publishing Group UK London}
+publisher={Nature Publishing Group UK London},
+doi={10.1038/s41467-021-23339-x},
 }
 @misc{wfacer,
 title = {WFacer},
@@ -60,7 +63,8 @@ volume={27},
 number={17},
 pages={5037--5059},
 year={2015},
-publisher={Wiley Online Library}
+publisher={Wiley Online Library},
+doi={10.1002/cpe.3505},
 }
 @article{da2023workflows,
 title={Workflows Community Summit 2022: A Roadmap Revolution},
@@ -74,7 +78,8 @@ author={Ben-Nun, Tal and Gamblin, Todd and Hollman, Daisy S and Krishnan, Hari a
 booktitle={2020 IEEE/ACM International Workshop on Performance, Portability and Productivity in HPC (P3HPC)},
 pages={57--69},
 year={2020},
-organization={IEEE}
+organization={IEEE},
+doi={10.1109/p3hpc51967.2020.00011},
 }
 @misc{wflowsystems,
 title={Existing Workflow Systems},
@@ -86,7 +91,8 @@ author={Al-Saadi, Aymen and Ahn, Dong H and Babuji, Yadu and Chard, Kyle and Cor
 booktitle={2021 IEEE Workshop on Workflows in Support of Large-Scale Science (WORKS)},
 pages={50--57},
 year={2021},
-organization={IEEE}
+organization={IEEE},
+doi={10.1109/works54523.2021.00012},
 }
 @inproceedings{babuji2019parsl,
 title={Parsl: Pervasive parallel programming in python},
@@ -107,7 +113,7 @@ url={https://github.com/insitro/redun}
 @misc{covalent,
 title={Covalent},
 year = {2023},
-doi={https://doi.org/10.5281/zenodo.5903364},
+doi={10.5281/zenodo.5903364},
 url={https://github.com/AgnostiqHQ/covalent}
 }
 @misc{maggma,
