### 2023-10-15 materialsproject/jobflow#453 @Andrew-S-Rosen
JOSS Paper: Ensure code blocks are presented nicely
As shown in [the draft](https://github.com/openjournals/joss-reviews/issues/5951#issuecomment-1763288768) of the Jobflow paper, one of the codeblock examples runs between pages. This is visually undesirable. For the final version, we will want to make sure this is avoided, either by adjusting where we place the code-block or by making a static image of the code block with smaller linespacing than JOSS auto-generates from Markdown.

![image](https://github.com/materialsproject/jobflow/assets/8674072/2136719f-c63c-436a-bedd-c6dcc83358b1)
