### 2023-10-15 #5951 @Andrew-S-Rosen
Here are some potential reviewers based on the JOSS reviewer list, where I considered individuals with expertise and interest in both chemistry/materials simulations and workflow infrastructure.

- Rasha Atwi (`rashatwi`)
- Javier Heras-Domingo (`jherasdo`)
- Alex Squires (`alexsquires`)
