### 2023-10-15 933342c
blacken-docs hook ignore paper/paper.md
