+++ .pre-commit-config.yaml
@@ -24,7 +24,7 @@ repos:
     hooks:
       - id: blacken-docs
         additional_dependencies: [black]
-        exclude: README.md
+        exclude: [README.md, paper/paper.md]
   - repo: https://github.com/pycqa/flake8
     rev: 6.0.0
     hooks:
