### 2023-10-15 d178f63
fix .pre-commit-config.yaml InvalidConfigError Hook(id='blacken-docs') exclude Expected string got list
