+++ .pre-commit-config.yaml
@@ -59,5 +59,5 @@ repos:
     hooks:
       - id: codespell
         stages: [commit, commit-msg]
-        args: [--ignore-words-list, "titel,statics,ba,nd,te"]
+        args: [--ignore-words-list, "titel,statics,ba,nd,te,atomate"]
         types_or: [python, rst, markdown]
