### 2023-10-16 materialsproject/jobflow#453 @utf
[paper-6.pdf](https://github.com/materialsproject/jobflow/files/12914011/paper-6.pdf)

Here is the latest paper pdf.
