### 2023-10-16 6e016b6
Bump moto from 4.2.5 to 4.2.6
Bumps [moto](https://github.com/getmoto/moto) from 4.2.5 to 4.2.6.
- [Release notes](https://github.com/getmoto/moto/releases)
- [Changelog](https://github.com/getmoto/moto/blob/master/CHANGELOG.md)
- [Commits](https://github.com/getmoto/moto/compare/4.2.5...4.2.6)

---
updated-dependencies:
- dependency-name: moto
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
