### 2023-10-16 3ec40cd
Bump maggma from 0.57.2 to 0.57.4 (#456)
Bumps [maggma](https://github.com/materialsproject/maggma) from 0.57.2 to 0.57.4.
- [Release notes](https://github.com/materialsproject/maggma/releases)
- [Changelog](https://github.com/materialsproject/maggma/blob/main/docs/CHANGELOG.md)
- [Commits](https://github.com/materialsproject/maggma/compare/v0.57.2...v0.57.4)

---
updated-dependencies:
- dependency-name: maggma
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+depend
[…]
