### 2023-10-17 e9989c1
Add missing year to `refs.bib`
