### 2023-10-17 0332851
JOSS Paper: Remove stray `
