### 2023-10-17 2224136
replace deprecated pkg_resources with importlib.metadata.version
