### 2023-10-17 b3eaa35
refactor tests/test_version.py
