### 2023-10-17 4f3c800
delete failing test_not_installed_version
