### 2023-10-17 8ed0367
coverage ignore PackageNotFoundError in _version.py
