+++ src/jobflow/_version.py
@@ -2,6 +2,6 @@ from importlib.metadata import PackageNotFoundError, version
 
 try:
     __version__ = version("jobflow")
-except PackageNotFoundError:
+except PackageNotFoundError:  # pragma: no cover
     # package is not installed
     __version__ = ""
