### 2023-10-21 ea75dfd
select same ruff rules as in atomate2
