### 2023-10-21 0cc30f9
fix boolean positional args
