### 2023-10-21 7189f02
remove unnecessary if/elif/else statements
