+++ src/jobflow/core/reference.py
@@ -161,9 +161,9 @@ class OutputReference(MSONable):
                 f"Could not resolve reference - {self.uuid}{istr} not in store or "
                 f"{index=}, {cache=}"
             )
-        elif on_missing == OnMissing.NONE and index not in cache[self.uuid]:
+        if on_missing == OnMissing.NONE and index not in cache[self.uuid]:
             return None
-        elif on_missing == OnMissing.PASS and index not in cache[self.uuid]:
+        if on_missing == OnMissing.PASS and index not in cache[self.uuid]:
             return self
 
         data = cache[self.uuid][index]
@@ -200,12 +200,11 @@ class OutputReference(MSONable):
         if inplace:
             self.uuid = uuid
             return self
-        else:
-            from copy import deepcopy
+        from copy import deepcopy
 
-            new_reference = deepcopy(self)
-            new_reference.uuid = uuid
-            return new_reference
+        new_reference = deepcopy(self)
+        new_reference.uuid = uuid
+        return new_reference
 
     def __getitem__(self, item) -> OutputReference:
         """Index the reference."""
@@ -285,7 +284,7 @@ class OutputReference(MSONable):
         """Serialize the reference as a dict."""
         schema = self.output_schema
         schema_dict = MontyEncoder().default(schema) if schema is not None else None
-        data = {
+        return {
             "@module": self.__class__.__module__,
             "@class": type(self).__name__,
             "@version": None,
@@ -293,7 +292,6 @@ class OutputReference(MSONable):
             "attributes": self.attributes,
             "output_schema": schema_dict,
         }
-        return data
 
 
 def resolve_references(
@@ -376,7 +374,7 @@ def find_and_get_references(arg: Any) -> tuple[OutputReference, ...]:
         # if the argument is a reference then stop there
         return (arg,)
 
-    elif isinstance(arg, (float, int, str, bool)):
+    if isinstance(arg, (float, int, str, bool)):
         # argument is a primitive, we won't find a reference here
         return ()
 
@@ -432,7 +430,7 @@ def find_and_resolve_references(
         # if the argument is a reference then stop there
         return arg.resolve(store, cache=cache, on_missing=on_missing)
 
-    elif isinstance(arg, (float, int, str, bool)):
+    if isinstance(arg, (float, int, str, bool)):
         # argument is a primitive, we won't find a reference here
         return arg
 
