+++ src/jobflow/core/store.py
@@ -250,8 +250,7 @@ class JobStore(Store):
         docs = self.query(
             criteria=criteria, properties=properties, load=load, sort=sort, limit=1
         )
-        d = next(docs, None)
-        return d
+        return next(docs, None)
 
     def update(
         self,
@@ -522,28 +521,27 @@ class JobStore(Store):
             return find_and_resolve_references(
                 result["output"], self, cache=cache, on_missing=on_missing
             )
-        else:
-            results = list(
-                self.query(
-                    criteria={"uuid": uuid},
-                    properties=["output"],
-                    sort={"index": 1},
-                    load=load,
-                )
+        results = list(
+            self.query(
+                criteria={"uuid": uuid},
+                properties=["output"],
+                sort={"index": 1},
+                load=load,
             )
+        )
 
-            if len(results) == 0:
-                raise ValueError(f"UUID: {uuid} has no outputs.")
+        if len(results) == 0:
+            raise ValueError(f"UUID: {uuid} has no outputs.")
 
-            results = [r["output"] for r in results]
+        results = [r["output"] for r in results]
 
-            refs = find_and_get_references(results)
-            if any(ref.uuid == uuid for ref in refs):
-                raise RuntimeError("Reference cycle detected - aborting.")
+        refs = find_and_get_references(results)
+        if any(ref.uuid == uuid for ref in refs):
+            raise RuntimeError("Reference cycle detected - aborting.")
 
-            return find_and_resolve_references(
-                results, self, cache=cache, on_missing=on_missing
-            )
+        return find_and_resolve_references(
+            results, self, cache=cache, on_missing=on_missing
+        )
 
     @classmethod
     def from_file(cls: type[T], db_file: str | Path, **kwargs) -> T:
