### 2023-10-21 bc31adf
move type imports into TYPE_CHECKING blocks
