+++ tests/managers/test_local.py
@@ -150,7 +150,7 @@ def test_detour_flow(memory_jobstore, clean_dir, detour_flow):
 
     # run with log
     responses = run_locally(flow, store=memory_jobstore)
-    uuid2 = next(u for u in responses if u != uuid1 and u != uuid3)
+    uuid2 = next(u for u in responses if u not in (uuid1, uuid3))
 
     # check responses has been filled
     assert len(responses) == 3
