+++ src/jobflow/managers/fireworks.py
@@ -10,6 +10,7 @@ if typing.TYPE_CHECKING:
     from collections.abc import Sequence
 
     import jobflow
+    from jobflow.core.job import Job
 
 
 def flow_to_workflow(
@@ -146,7 +147,6 @@ class JobFiretask(FiretaskBase):
     def run_task(self, fw_spec):
         """Run the job and handle any dynamic firework submissions."""
         from jobflow import SETTINGS, initialize_logger
-        from jobflow.core.job import Job
 
         job: Job = self.get("job")
         store = self.get("store")
