### 2023-10-21 a34e772
fix ruff PLR6201
Use a set literal when testing for membership
