### 2023-10-22 #5951 @arfon
@jjmortensen @paulromano @hjoliver @rashatwi @jherasdo – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is Jobflow: Computational Workflows Made Simple: https://github.com/openjournals/joss-reviews/issues/5951

Based on your experience and previous publications in JOSS, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
