### 2023-10-22 fc28f75
fix all ruff B018 and PT011 violations
