### 2023-10-22 ab71d1d
remove ruff.per-file-ignores for B018 and PT011 in tests
