+++ pyproject.toml
@@ -151,16 +151,12 @@ select = [
     "W",      # pycodestyle warning
     "YTT",    # flake8-2020
 ]
-# PLR0911: too-many-return-statements
-# PLR0912: too-many-branches
-# PLR0913: too-many-arguments
-# PLR0915: too-many-statements
 ignore = [
     "B028",
-    "PLR0911",
-    "PLR0912",
-    "PLR0913",
-    "PLR0915",
+    "PLR0911", # too-many-return-statements
+    "PLR0912", # too-many-branches
+    "PLR0913", # too-many-arguments
+    "PLR0915", # too-many-statements
     "PLW0603",
     "RUF013",
 ]
@@ -168,9 +164,9 @@ pydocstyle.convention = "numpy"
 isort.known-first-party = ["jobflow"]
 
 [tool.ruff.per-file-ignores]
+# F401: unused import
 "__init__.py" = ["F401"]
+# D: pydocstyle
 # PLR2004: magic-value-comparison
 # PT004: pytest-missing-fixture-name-underscore
-# PLR0915: too-many-statements
-# PT011: pytest-raises-too-broad TODO fix these, should not be ignored
-"**/tests/*" = ["B018", "D", "PLR0915", "PLR2004", "PT004", "PT011"]
+"**/tests/*" = ["D", "PLR2004", "PT004"]
