### 2023-10-23 #5951 @jherasdo
Hi Arfon!

Thank you so much for consider me as a reviewer for JOSS!

I do accept to be reviewer for this publication.

I have some questions since this is my first time being reviewer for your
journal:

- There is any deadline to submit my revisions?
- I still need to read the guidelines that you provided, but could send me
an example of peer-review for another article? (Just to see the whole
process beforehand)

Let me know if you need anything else from my side!

Thank you so much!
Dr. Javier Heras-Domingo

Missatge de Arfon Smith ***@***.***> del dia dg., 22 d’oct.
2023 a les 13:07:

> @jjmortensen <https://github.com/jjmortensen> @paulromano
> <https://github.com/paulromano> @hjoliver <https://github.com/hjoliver>
> @rashatwi <https://github.com/rashatwi> @jherasdo
> <https://github.com/jherasdo> – 👋 would any of you be willing to review
> this submission for JOSS? The submission under consideration is Jobflow:
> Computational Workflows Made Simple: #5951
> <https://github.com/openjournals/joss-reviews/issues/5951>
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. You can
> learn more about the process in these guidelines:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> Based on your experience and previous publications in JOSS, we think you
> might be able to provide a great review of this submission. Please let me
> know if you think you can help us out!
>
> Many thanks
> Arfon
>
