### 2023-10-23 0c01656
Bump networkx from 3.1 to 3.2 (#466)
Bumps [networkx](https://github.com/networkx/networkx) from 3.1 to 3.2.
- [Release notes](https://github.com/networkx/networkx/releases)
- [Commits](https://github.com/networkx/networkx/compare/networkx-3.1...networkx-3.2)

---
updated-dependencies:
- dependency-name: networkx
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
