### 2023-10-23 #5951 @arfon
> I do accept to be reviewer for this publication.

Great stuff, thanks!

> - There is any deadline to submit my revisions?

We generally ask for reviews to be completed in 4-6 weeks. As JOSS reviews are conversational in nature, it's a good idea to start sooner than that though.

> - I still need to read the guidelines that you provided, but could send me
> an example of peer-review for another article? (Just to see the whole
> process beforehand)

Certainly! Here's a review for a recently-accepted paper in JOSS: https://github.com/openjournals/joss-reviews/issues/5776
