### 2023-10-29 8b6dca2
simplify tests by using flow __len__ and __getitem__ methods
