+++ tests/core/test_flow.py
@@ -149,7 +149,7 @@ def test_flow_of_flows_init():
     assert flow.output is None
     assert flow.job_uuids == (add_job.uuid,)
     assert flow.all_uuids == (add_job.uuid, sub_flow.uuid)
-    assert flow.jobs[0].host == flow.uuid
+    assert flow[0].host == flow.uuid
 
     # test single flow no list
     add_job = get_test_job()
@@ -159,7 +159,7 @@ def test_flow_of_flows_init():
     assert flow.host is None
     assert flow.output is None
     assert flow.job_uuids == (add_job.uuid,)
-    assert flow.jobs[0].host == flow.uuid
+    assert flow[0].host == flow.uuid
 
     # test multiple flows
     add_job1 = get_test_job()
@@ -176,8 +176,8 @@ def test_flow_of_flows_init():
         add_job2.uuid,
         subflow2.uuid,
     )
-    assert flow.jobs[0].host == flow.uuid
-    assert flow.jobs[1].host == flow.uuid
+    assert flow[0].host == flow.uuid
+    assert flow[1].host == flow.uuid
 
     # test single job and outputs
     add_job = get_test_job()
@@ -237,8 +237,8 @@ def test_flow_job_mixed():
     assert flow.host is None
     assert flow.output is None
     assert flow.job_uuids == (add_job.uuid, add_job2.uuid)
-    assert flow.jobs[0].host == flow.uuid
-    assert flow.jobs[1].host == flow.uuid
+    assert flow[0].host == flow.uuid
+    assert flow[1].host == flow.uuid
 
     # test with list multi outputs
     add_job = get_test_job()
@@ -496,41 +496,41 @@ def test_serialization():
     encoded_flow = json.loads(MontyEncoder().encode(flow_host))
     decoded_flow = MontyDecoder().process_decoded(encoded_flow)
 
-    assert decoded_flow.jobs[0].host == host_uuid
+    assert decoded_flow[0].host == host_uuid
 
 
 def test_update_kwargs():
     # test no filter
     flow = get_test_flow()
     flow.update_kwargs({"b": 5})
-    assert flow.jobs[0].function_kwargs["b"] == 5
-    assert flow.jobs[1].function_kwargs["b"] == 5
+    assert flow[0].function_kwargs["b"] == 5
+    assert flow[1].function_kwargs["b"] == 5
 
     # test name filter
     flow = get_test_flow()
     flow.update_kwargs({"b": 5}, name_filter="div")
-    assert "b" not in flow.jobs[0].function_kwargs
-    assert flow.jobs[1].function_kwargs["b"] == 5
+    assert "b" not in flow[0].function_kwargs
+    assert flow[1].function_kwargs["b"] == 5
 
     # test function filter
     flow = get_test_flow()
     flow.update_kwargs({"b": 5}, function_filter=div)
-    assert "b" not in flow.jobs[0].function_kwargs
-    assert flow.jobs[1].function_kwargs["b"] == 5
+    assert "b" not in flow[0].function_kwargs
+    assert flow[1].function_kwargs["b"] == 5
 
     # test dict mod
     flow = get_test_flow()
     flow.update_kwargs({"_inc": {"b": 5}}, function_filter=div, dict_mod=True)
-    assert "b" not in flow.jobs[0].function_kwargs
-    assert flow.jobs[1].function_kwargs["b"] == 8
+    assert "b" not in flow[0].function_kwargs
+    assert flow[1].function_kwargs["b"] == 8
 
 
 def test_update_maker_kwargs():
     # test no filter
     flow = get_maker_flow()
     flow.update_maker_kwargs({"b": 10})
-    assert flow.jobs[0].maker.b == 10
[…]
