+++ src/jobflow/core/flow.py
@@ -166,7 +166,7 @@ class Flow(MSONable):
             raise TypeError(
                 f"Flow can only contain Job or Flow objects, not {type(value).__name__}"
             )
-        jobs = list(self.jobs)
+        jobs = list(self)
         jobs[idx] = value  # type: ignore[index, assignment]
         self.jobs = tuple(jobs)
 
@@ -176,7 +176,7 @@ class Flow(MSONable):
 
     def __contains__(self, item: Flow | Job) -> bool:
         """Check if the flow contains a job or subflow."""
-        return item in self.jobs
+        return item in self
 
     def __add__(self, other: Job | Flow | Sequence[Flow | Job]) -> Flow:
         """Add a job or subflow to the flow."""
@@ -188,7 +188,7 @@ class Flow(MSONable):
 
     def __sub__(self, other: Flow | Job) -> Flow:
         """Remove a job or subflow from the flow."""
-        if other not in self.jobs:
+        if other not in self:
             raise ValueError(f"{other!r} not found in flow")
         new_flow = deepcopy(self)
         new_flow.jobs = tuple([job for job in new_flow if job != other])
@@ -201,8 +201,8 @@ class Flow(MSONable):
         _prefix = f"{prefix}." if prefix else ""
         job_reprs = "\n".join(
             f"{indent}{_prefix}{i}. "
-            f"{j.__repr__(level + 1, f'{_prefix}{i}') if isinstance(j, Flow) else j}"
-            for i, j in enumerate(self.jobs, 1)
+            f"{jb.__repr__(level + 1, f'{_prefix}{i}') if isinstance(jb, Flow) else jb}"
+            for i, jb in enumerate(self, 1)
         )
         return f"Flow({name=}, {uuid=})\n{job_reprs}"
 
@@ -298,7 +298,7 @@ class Flow(MSONable):
             The uuids of all Jobs in the Flow (including nested Flows).
         """
         uuids: list[str] = []
-        for job in self.jobs:
+        for job in self:
             if isinstance(job, Flow):
                 uuids.extend(job.job_uuids)
             else:
@@ -316,7 +316,7 @@ class Flow(MSONable):
             The uuids of all Jobs and Flows in the Flow (including nested Flows).
         """
         uuids: list[str] = []
-        for job in self.jobs:
+        for job in self:
             if isinstance(job, Flow):
                 uuids.extend(job.all_uuids)
             uuids.append(job.uuid)
@@ -336,7 +336,7 @@ class Flow(MSONable):
 
         import networkx as nx
 
-        graph = nx.compose_all([job.graph for job in self.jobs])
+        graph = nx.compose_all([job.graph for job in self])
 
         for node in graph:
             node_props = graph.nodes[node]
@@ -346,7 +346,7 @@ class Flow(MSONable):
         if self.order == JobOrder.LINEAR:
             # add fake edges between jobs to force linear order
             edges = []
-            for job_a, job_b in nx.utils.pairwise(self.jobs):
+            for job_a, job_b in nx.utils.pairwise(self):
                 if isinstance(job_a, Flow):
                     leaves = [v for v, d in job_a.graph.out_degree() if d == 0]
                 else:
@@ -474,7 +474,7 @@ class Flow(MSONable):
         >>> flow.update_kwargs({"number": 10}, name_filter="add")
         >>> flow.update_kwargs({"number": 10}, function_filter=add)
         """
-        for job in self.jobs:
+        for job in self:
             job.update_kwargs(
                 update,
                 name_filter=name_filter,
@@ -573,7 +573,7 @@ class Flow(MSONable):
         ...     {"number": 10}, class_filter=AddMaker, nested=False
         ... )
         """
-        for job in self.jobs:
+        for job in self:
[…]
