+++ tests/core/test_job.py
@@ -161,23 +161,23 @@ def test_replace_response(memory_jobstore):
     test_job = Job(replace_job, metadata=metadata, output_schema="123")
     response = test_job.run(memory_jobstore)
     assert isinstance(response.replace, Flow)
-    assert len(response.replace.jobs) == 1
-    assert response.replace.jobs[0].index == 2
-    assert response.replace.jobs[0].uuid == test_job.uuid
-    assert response.replace.jobs[0].metadata == metadata
-    assert response.replace.jobs[0].output_schema == "123"
+    assert len(response.replace) == 1
+    assert response.replace[0].index == 2
+    assert response.replace[0].uuid == test_job.uuid
+    assert response.replace[0].metadata == metadata
+    assert response.replace[0].output_schema == "123"
     assert test_job.hosts == []
     assert response.replace.hosts == []
-    assert response.replace.jobs[0].hosts == [response.replace.uuid]
+    assert response.replace[0].hosts == [response.replace.uuid]
 
     # replace with list of job
     test_job = Job(replace_list_job, metadata=metadata, output_schema="123")
     response = test_job.run(memory_jobstore)
     assert isinstance(response.replace, Flow)
-    assert response.replace.jobs[-1].function == add
-    assert len(response.replace.jobs) == 2
+    assert response.replace[-1].function == add
+    assert len(response.replace) == 2
     # currently output schema and metadata ignored in this case
-    for j in response.replace.jobs:
+    for j in response.replace:
         assert j.hosts == [response.replace.uuid]
 
     # replace with flow with outputs
@@ -186,33 +186,33 @@ def test_replace_response(memory_jobstore):
     test_flow = Flow([test_job])
     response = test_job.run(memory_jobstore)
     assert isinstance(response.replace, Flow)
-    assert response.replace.jobs[-1].index == 2
-    assert response.replace.jobs[-1].uuid == test_job.uuid
-    assert response.replace.jobs[-1].metadata == metadata
-    assert response.replace.jobs[-1].output_schema == "123"
-    assert response.replace.jobs[-1]._kwargs["data"]
+    assert response.replace[-1].index == 2
+    assert response.replace[-1].uuid == test_job.uuid
+    assert response.replace[-1].metadata == metadata
+    assert response.replace[-1].output_schema == "123"
+    assert response.replace[-1]._kwargs["data"]
     assert response.replace.output is not None
-    for j in response.replace.jobs:
+    for j in response.replace:
         assert j.hosts == [response.replace.uuid, test_flow.uuid]
 
     # replace with flow with multi outputs
     test_job = Job(replace_flow_multioutput, metadata=metadata, output_schema="123")
     response = test_job.run(memory_jobstore)
     assert isinstance(response.replace, Flow)
-    assert response.replace.jobs[-1].index == 2
-    assert response.replace.jobs[-1].uuid == test_job.uuid
-    assert response.replace.jobs[-1].metadata == metadata
-    assert response.replace.jobs[-1].output_schema == "123"
+    assert response.replace[-1].index == 2
+    assert response.replace[-1].uuid == test_job.uuid
+    assert response.replace[-1].metadata == metadata
+    assert response.replace[-1].output_schema == "123"
     assert response.replace.output is not None
 
     # replace with list of flow
     test_job = Job(replace_list_flow, metadata=metadata, output_schema="123")
     response = test_job.run(memory_jobstore)
     assert isinstance(response.replace, Flow)
-    assert isinstance(response.replace.jobs[-1], Flow)
-    assert len(response.replace.jobs) == 2
-    for f in response.replace.jobs:
-        for j in f.jobs:
+    assert isinstance(response.replace[-1], Flow)
+    assert len(response.replace) == 2
+    for f in response.replace:
[…]
