### 2023-10-29 7c66385
fix test_flow.py::test_flow_magic_methods - RecursionError: maximum recursion depth exceeded in __contains__
