### 2023-10-30 6ab8d8a
Bump ipython from 8.16.1 to 8.17.1
Bumps [ipython](https://github.com/ipython/ipython) from 8.16.1 to 8.17.1.
- [Release notes](https://github.com/ipython/ipython/releases)
- [Commits](https://github.com/ipython/ipython/commits/8.17.1)

---
updated-dependencies:
- dependency-name: ipython
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
