+++ pyproject.toml
@@ -39,7 +39,7 @@ dependencies = [
 docs = [
     "autodoc_pydantic==2.0.1",
     "furo==2023.9.10",
-    "ipython==8.16.1",
+    "ipython==8.17.1",
     "myst_parser==2.0.0",
     "nbsphinx==0.9.3",
     "sphinx-copybutton==0.5.2",
