+++ pyproject.toml
@@ -46,7 +46,7 @@ docs = [
     "sphinx==7.2.6",
 ]
 dev = ["pre-commit>=2.12.1"]
-tests = ["moto==4.2.6", "pytest-cov==4.1.0", "pytest==7.4.2"]
+tests = ["moto==4.2.6", "pytest-cov==4.1.0", "pytest==7.4.3"]
 vis = ["matplotlib", "pydot"]
 fireworks = ["FireWorks"]
 strict = [
