### 2023-10-30 4a848bb
Bump moto from 4.2.6 to 4.2.7 (#470)
Bumps [moto](https://github.com/getmoto/moto) from 4.2.6 to 4.2.7.
- [Release notes](https://github.com/getmoto/moto/releases)
- [Changelog](https://github.com/getmoto/moto/blob/master/CHANGELOG.md)
- [Commits](https://github.com/getmoto/moto/compare/4.2.6...4.2.7)

---
updated-dependencies:
- dependency-name: moto
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
