+++ .pre-commit-config.yaml
@@ -3,10 +3,11 @@ default_language_version:
 exclude: "^src/atomate2/vasp/schemas/calc_types/"
 repos:
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.1.1
+    rev: v0.1.3
     hooks:
       - id: ruff
         args: [--fix]
+      - id: ruff-format
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.5.0
     hooks:
@@ -15,10 +16,6 @@ repos:
         args: [--remove]
       - id: end-of-file-fixer
       - id: trailing-whitespace
-  - repo: https://github.com/psf/black
-    rev: 23.10.0
-    hooks:
-      - id: black
   - repo: https://github.com/asottile/blacken-docs
     rev: 1.16.0
     hooks:
@@ -59,5 +56,8 @@ repos:
     hooks:
       - id: codespell
         stages: [commit, commit-msg]
-        args: [--ignore-words-list, "titel,statics,ba,nd,te,atomate"]
+        args:
+          - --ignore-words-list
+          - "titel,statics,ba,nd,te,atomate"
+          - --check-filenames
         types_or: [python, rst, markdown]
