### 2023-11-06 6adf959
Bump matplotlib from 3.7.2 to 3.8.1
Bumps [matplotlib](https://github.com/matplotlib/matplotlib) from 3.7.2 to 3.8.1.
- [Release notes](https://github.com/matplotlib/matplotlib/releases)
- [Commits](https://github.com/matplotlib/matplotlib/compare/v3.7.2...v3.8.1)

---
updated-dependencies:
- dependency-name: matplotlib
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
