+++ pyproject.toml
@@ -53,7 +53,7 @@ strict = [
     "FireWorks==2.0.3",
     "PyYAML==6.0.1",
     "maggma==0.57.4",
-    "matplotlib==3.7.2",
+    "matplotlib==3.8.1",
     "monty==2023.9.25",
     "moto==4.2.7",
     "networkx==3.2.1",
