### 2023-11-06 792c4c1
Bump ipython from 8.17.1 to 8.17.2
Bumps [ipython](https://github.com/ipython/ipython) from 8.17.1 to 8.17.2.
- [Release notes](https://github.com/ipython/ipython/releases)
- [Commits](https://github.com/ipython/ipython/compare/8.17.1...8.17.2)

---
updated-dependencies:
- dependency-name: ipython
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
