### 2023-11-06 aad64a7
Bump maggma from 0.57.4 to 0.57.5 (#474)
Bumps [maggma](https://github.com/materialsproject/maggma) from 0.57.4 to 0.57.5.
- [Release notes](https://github.com/materialsproject/maggma/releases)
- [Changelog](https://github.com/materialsproject/maggma/blob/main/docs/CHANGELOG.md)
- [Commits](https://github.com/materialsproject/maggma/compare/v0.57.4...v0.57.5)

---
updated-dependencies:
- dependency-name: maggma
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+depend
[…]
