+++ pyproject.toml
@@ -52,7 +52,7 @@ fireworks = ["FireWorks"]
 strict = [
     "FireWorks==2.0.3",
     "PyYAML==6.0.1",
-    "maggma==0.57.4",
+    "maggma==0.57.5",
     "matplotlib==3.8.1",
     "monty==2023.9.25",
     "moto==4.2.7",
