### 2023-11-06 92058b6
Bump monty from 2023.9.25 to 2023.11.3 (#475)
Bumps [monty](https://github.com/materialsvirtuallab/monty) from 2023.9.25 to 2023.11.3.
- [Release notes](https://github.com/materialsvirtuallab/monty/releases)
- [Changelog](https://github.com/materialsvirtuallab/monty/blob/master/docs/changelog.md)
- [Commits](https://github.com/materialsvirtuallab/monty/compare/v2023.9.25...v2023.11.3)

---
updated-dependencies:
- dependency-name: monty
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependa
[…]
