### 2023-11-08 66296ea
Fix JobStoreDocument validator
