### 2023-11-13 #5995 @jherasdo
The manuscript “Jobflow: Computational Workflows Made Simple” by Andrew S. Rosen and co-workers present the software application called Jobflow as an open source Python package for writing computational workflows tailored for high-throughput computing applications. This application abstracts a lot of complexities in the field by using a decorator-based approach which is able to transform pythonic functions and classes into more complex workflows. Furthermore, Jobflow also supports dynamic workflows and different ways to store the results both locally and cloud-based databases. Finally, this application can be coupled with FireWorks package to easily manage computational resources in typical HPC cluster architectures.

I have completed my first review of the code and the paper. The software looks very good. The paper is well written with a clear statement of need and state of the field. There are only few things to solve in the manuscript:

- There is a minor typo in line 80: A missing comma after e.g. (for consistency with the other “e.g.”, I would add it).

I encourage the authors to revise the references since there are some issues:
- It is a bit weird the use of “& others” since there is more than enough space to include all authors in each reference.
- For the Covalent software package reference, the authors provide the zenodo link but didn’t include any author or the format of citation suggested in the zenodo link.
- For the reference “Existing workflow systems. (2023).” The README file from the github has a different way of citing this work. Seems like should be cited through the following work https://dl.acm.org/doi/10.1145/3486897
