### 2023-11-13 63ca365
JOSS: Add missing `,` to paper
Addressed first bullet point by reviewer: https://github.com/openjournals/joss-reviews/issues/5995#issuecomment-1808396589
