### 2023-11-13 3b47876
JOSS: Fix references to include all coauthors and remove "others"
This was a leftover issue from fetching references from Google Scholar.

Addresses https://github.com/openjournals/joss-reviews/issues/5995#issuecomment-1808396589
