+++ paper/refs.bib
@@ -11,7 +11,7 @@ url = {https://github.com/Quantum-Accelerators/quacc}
 }
 @article{materialsproject,
   title={Commentary: The Materials Project: A materials genome approach to accelerating materials innovation},
-  author={Jain, Anubhav and Ong, Shyue Ping and Hautier, Geoffroy and Chen, Wei and Richards, William Davidson and Dacek, Stephen and Cholia, Shreyas and Gunter, Dan and Skinner, David and Ceder, Gerbrand and others},
+  author={Jain, Anubhav and Ong, Shyue Ping and Hautier, Geoffroy and Chen, Wei and Richards, William Davidson and Dacek, Stephen and Cholia, Shreyas and Gunter, Dan and Skinner, David and Ceder, Gerbrand and Persson, Kristin A},
   journal={APL materials},
   volume={1},
   number={1},
@@ -57,7 +57,7 @@ url = {https://github.com/BlauGroup/NanoParticleTools}
 }
 @article{fireworks,
 title={FireWorks: a dynamic workflow system designed for high-throughput applications},
-author={Jain, Anubhav and Ong, Shyue Ping and Chen, Wei and Medasani, Bharat and Qu, Xiaohui and Kocher, Michael and Brafman, Miriam and Petretto, Guido and Rignanese, Gian-Marco and Hautier, Geoffroy and others},
+author={Jain, Anubhav and Ong, Shyue Ping and Chen, Wei and Medasani, Bharat and Qu, Xiaohui and Kocher, Michael and Brafman, Miriam and Petretto, Guido and Rignanese, Gian-Marco and Hautier, Geoffroy and Gunter, Daniel and Persson, Kristin A},
 journal={Concurrency and Computation: Practice and Experience},
 volume={27},
 number={17},
@@ -68,7 +68,7 @@ doi={10.1002/cpe.3505},
 }
 @article{da2023workflows,
 title={Workflows Community Summit 2022: A Roadmap Revolution},
-author={da Silva, Rafael Ferreira and Badia, Rosa M and Bala, Venkat and Bard, Debbie and Bremer, Peer-Timo and Buckley, Ian and Caino-Lores, Silvina and Chard, Kyle and Goble, Carole and Jha, Shantenu and others},
[…]
