+++ paper/refs.bib
@@ -8,9 +8,8 @@ url = {https://github.com/materialsproject/atomate2}
   title        = {quacc – The Quantum Accelerator},
   year         = 2023,
   publisher    = {Zenodo},
-  version      = {v0.4.0},
-  doi          = {10.5281/zenodo.10091610},
-  url          = {https://doi.org/10.5281/zenodo.10091610}
+  doi          = {10.5281/zenodo.7720998},
+  url          = {https://doi.org/10.5281/zenodo.7720998}
 }
 @article{materialsproject,
   title={Commentary: The Materials Project: A materials genome approach to accelerating materials innovation},
