+++ paper/refs.bib
@@ -146,9 +146,8 @@ url={https://github.com/insitro/redun}
   title        = {Covalent},
   year         = 2023,
   publisher    = {Zenodo},
-  version      = {v0.228.0-rc.0},
-  doi          = {10.5281/zenodo.8369670},
-  url          = {https://doi.org/10.5281/zenodo.8369670}
+  doi          = {10.5281/zenodo.5903364},
+  url          = {https://doi.org/10.5281/zenodo.5903364}
 }
 @misc{maggma,
 title={Maggma},
