+++ paper/refs.bib
@@ -6,7 +6,7 @@ url = {https://github.com/materialsproject/atomate2}
 @software{quacc,
   author       = {Rosen, Andrew S},
   title        = {quacc – The Quantum Accelerator},
-  year         = 2023,
+  year         = {2023},
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.7720998},
   url          = {https://doi.org/10.5281/zenodo.7720998}
@@ -144,7 +144,7 @@ url={https://github.com/insitro/redun}
                   Singh, Divyanshu and
                   FilipBolt},
   title        = {Covalent},
-  year         = 2023,
+  year         = {2023},
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.5903364},
   url          = {https://doi.org/10.5281/zenodo.5903364}
