### 2023-11-13 #5995 @Andrew-S-Rosen
@jherasdo: Thank you for your positive feedback and, of course, for your review! We greatly appreciate the time you spent on this. I'll briefly address your comments, but please let us know if you have any further requests or comments!

> There is a minor typo in line 80: A missing comma after e.g. (for consistency with the other “e.g.”, I would add it).

Thanks! Fixed in the updated paper.

> It is a bit weird the use of “& others” since there is more than enough space to include all authors in each reference.

Thanks for catching this. It is now fixed. In case you're curious, this was from Google Scholar's bibtex exporter, which automatically truncates author lists with "& others." Very annoying. I must have missed it.

> For the Covalent software package reference, the authors provide the zenodo link but didn’t include any author or the format of citation suggested in the zenodo link.

Thanks! This is now fixed (and I did the same for the `quacc` package, which also had a Zenodo DOI).

> For the reference “Existing workflow systems. (2023).” The README file from the github has a different way of citing this work. Seems like should be cited through the following work https://dl.acm.org/doi/10.1145/3486897

Thanks. The reference you are referring to is slightly different. The [list of existing workflow systems](https://github.com/common-workflow-language/common-workflow-language/wiki/Existing-Workflow-systems) happens to be hosted on the CWL (common workflow language) repository but is a distinct entity from CWL itself. The reference you shared is specifically for CWL, which is a bit outside the scope of this manuscript.
