### 2023-11-13 a9c44db
Bump maggma from 0.57.5 to 0.57.8
Bumps [maggma](https://github.com/materialsproject/maggma) from 0.57.5 to 0.57.8.
- [Release notes](https://github.com/materialsproject/maggma/releases)
- [Changelog](https://github.com/materialsproject/maggma/blob/main/docs/CHANGELOG.md)
- [Commits](https://github.com/materialsproject/maggma/compare/v0.57.5...v0.57.8)

---
updated-dependencies:
- dependency-name: maggma
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
