+++ pyproject.toml
@@ -52,7 +52,7 @@ fireworks = ["FireWorks"]
 strict = [
     "FireWorks==2.0.3",
     "PyYAML==6.0.1",
-    "maggma==0.57.5",
+    "maggma==0.57.8",
     "matplotlib==3.8.1",
     "monty==2023.11.3",
     "moto==4.2.8",
