### 2023-11-14 #5995 @jherasdo
Thank you so much for your quick answer @Andrew-S-Rosen! I have updated my checklist accordingly!

I actually have a couple of questions regarding Jobflow:

Since I'm more used to interact with the "original" atomate package.

- I'm curious to know if the Jobflow/atomate2 combination is going to be able to take complex data as input and to serialize it as in atomate and what part of this is abstracted inside Joblow.

And also, since sometimes is quite difficult to deploy and maintain a MongoDB in HPC clusters:

- Are you "The Jobflow Team" thinking on also implementing the possibility to use PostgresSQL as "local" database, similar to other workflow managers like AiiDa?

Thanks again!
