### 2023-11-15 #5995 @jherasdo
Thank you so much @Andrew-S-Rosen and @utf for your answers and congrats for the nice work!

I will definitely take a look on the DB possibilities you mentioned. And also thanks for considering Postgres as a future addition for maggma and Joblow.

@arfon from my side everything looks good and my questions/comments were solved accordingly.

Please let me know if you need anything else from my side.
