### 2023-11-16 bf36d98
test run_locally with root_dir
