### 2023-11-21 658ce76
Bump moto from 4.2.8 to 4.2.9
Bumps [moto](https://github.com/getmoto/moto) from 4.2.8 to 4.2.9.
- [Release notes](https://github.com/getmoto/moto/releases)
- [Changelog](https://github.com/getmoto/moto/blob/master/CHANGELOG.md)
- [Commits](https://github.com/getmoto/moto/compare/4.2.8...4.2.9)

---
updated-dependencies:
- dependency-name: moto
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
