### 2023-11-21 0d9ae3e
Bump matplotlib from 3.8.1 to 3.8.2
Bumps [matplotlib](https://github.com/matplotlib/matplotlib) from 3.8.1 to 3.8.2.
- [Release notes](https://github.com/matplotlib/matplotlib/releases)
- [Commits](https://github.com/matplotlib/matplotlib/compare/v3.8.1...v3.8.2)

---
updated-dependencies:
- dependency-name: matplotlib
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
