+++ pyproject.toml
@@ -53,7 +53,7 @@ strict = [
     "FireWorks==2.0.3",
     "PyYAML==6.0.1",
     "maggma==0.57.10",
-    "matplotlib==3.8.1",
+    "matplotlib==3.8.2",
     "monty==2023.11.3",
     "moto==4.2.9",
     "networkx==3.2.1",
