### 2023-11-27 2740742
Bump moto from 4.2.9 to 4.2.10
Bumps [moto](https://github.com/getmoto/moto) from 4.2.9 to 4.2.10.
- [Release notes](https://github.com/getmoto/moto/releases)
- [Changelog](https://github.com/getmoto/moto/blob/master/CHANGELOG.md)
- [Commits](https://github.com/getmoto/moto/compare/4.2.9...4.2.10)

---
updated-dependencies:
- dependency-name: moto
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
