+++ pyproject.toml
@@ -58,7 +58,7 @@ strict = [
     "moto==4.2.9",
     "networkx==3.2.1",
     "pydantic-settings==2.0.3",
-    "pydantic==2.5.1",
+    "pydantic==2.5.2",
     "pydash==7.0.6",
     "pydot==1.4.2",
     "typing-extensions==4.8.0",
