### 2023-11-27 3b18088
Bump maggma from 0.57.10 to 0.58.0 (#498)
Bumps [maggma](https://github.com/materialsproject/maggma) from 0.57.10 to 0.58.0.
- [Release notes](https://github.com/materialsproject/maggma/releases)
- [Changelog](https://github.com/materialsproject/maggma/blob/main/docs/CHANGELOG.md)
- [Commits](https://github.com/materialsproject/maggma/compare/v0.57.10...v0.58.0)

---
updated-dependencies:
- dependency-name: maggma
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dep
[…]
