### 2023-11-27 a29ad52
Bump pydantic-settings from 2.0.3 to 2.1.0
Bumps [pydantic-settings](https://github.com/pydantic/pydantic-settings) from 2.0.3 to 2.1.0.
- [Release notes](https://github.com/pydantic/pydantic-settings/releases)
- [Commits](https://github.com/pydantic/pydantic-settings/compare/v2.0.3...v2.1.0)

---
updated-dependencies:
- dependency-name: pydantic-settings
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
