+++ pyproject.toml
@@ -57,7 +57,7 @@ strict = [
     "monty==2023.11.3",
     "moto==4.2.10",
     "networkx==3.2.1",
-    "pydantic-settings==2.0.3",
+    "pydantic-settings==2.1.0",
     "pydantic==2.5.2",
     "pydash==7.0.6",
     "pydot==1.4.2",
