### 2023-12-04 0933fbb
Bump moto from 4.2.10 to 4.2.11
Bumps [moto](https://github.com/getmoto/moto) from 4.2.10 to 4.2.11.
- [Release notes](https://github.com/getmoto/moto/releases)
- [Changelog](https://github.com/getmoto/moto/blob/master/CHANGELOG.md)
- [Commits](https://github.com/getmoto/moto/compare/4.2.10...4.2.11)

---
updated-dependencies:
- dependency-name: moto
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
