### 2023-12-06 d9ea096
fix pass manager config for dynamic jobs
