+++ src/jobflow/core/job.py
@@ -1347,6 +1347,7 @@ def pass_manager_config(
 ):
     """
     Pass the manager config on to any jobs in the jobs array.
+    Merge with already specified manager config.
 
     Parameters
     ----------
