### 2023-12-08 #5995 @rashatwi
# Overview
I finished my review of the paper and the code. The Jobflow package is well suited for JOSS and is of significant interest to the computational science community. I was really impressed by how intuitive it seems to construct Flows using Jobflow. Some of the features I really like are how the output of each job is stored, how the dependency between Jobs is automatically determined based on inputs, and how you can simply register the output in the Flow object.

As a user of the FireWorks package, I really appreciate the flexibility that Jobflow offers in terms of choosing the workflow manager and database as well as the short overhead time it seems to take to get started writing Flows.  I found that some of the design concepts/features are common with FireWorks, e.g., dynamic workflows, detours, and ability to modify parameters of a Job after it has been defined, which are all desirable if someone is used to FireWorks and doesn’t want to lose access to these features when making the transition.

I don’t have comments related to the paper, as it is very clear and the examples provided are simple and straight to the point. I also checked off most of the items on the checklist, but I have some minor comments related to the documentation.

# Documentation

1. The link in this part of the documentation **(Defining jobs in jobflow)** is not working. Perhaps you were trying to refer to [this](https://materialsproject.github.io/jobflow/tutorials/5-dynamic-flows.html) section instead:

> Jobs also have an index. This tracks the number of times the job has been “replaced” (replacing is covered in detail in the [[Dynamic and nested Flows tutorial](https://materialsproject.github.io/jobflow/tutorials/dynamic-flows)].

2. **Introduction to Jobflow tutorial**: After executing
`flow = Flow(
    jobs=[time_github, time_google, time_nyt, sum_times],
    output=sum_times.output
)`
I got the following error:

> ValueError: Job time_website (cba5389e-8837-4bf1-aee1-87b650192bc1) already belongs to another flow.

3. I found the tutorials provided very basic and some are repetitive, which might be intentional. However, I would appreciate the addition of a bit more advanced examples, e.g., there is not much about setting a custom store other than `MemoryStore`. The docstrings provided are a very good starting point, but I’m not sure if I’m missing any additional information out there about the custom stores.

4. One of the main features of Jobflow is being independent of the workflow manager and you already have an adapter/support for FireWorks. I was wondering if you are planning to add additional adapters, and if so, which ones? If not and if this is a task left to the user to create their own adapters, it would be nice to have an example/guide on the website on how to get started.

# General Questions
Aside from the review, I have the following questions:

1. Is there support in the Job output for a wider range of datatypes, not only limited to Monty serializable datatypes?

2. Having rewritten atomate2 based on atomate, how difficult was the transition from FireWorks to Jobflow for defining the flows, i.e., what was the overhead time for modifying/building upon existing workflows that utilize the FireWorks structure to the use Jobflow instead. I understand that this is a very subjective question, but any comment about that would be really appreciated. I would imagine there are lots of people who have been using the FireWorks package for defining workflows and might be interested in making this transition as well.
