+++ .pre-commit-config.yaml
@@ -3,7 +3,7 @@ default_language_version:
 exclude: "^src/atomate2/vasp/schemas/calc_types/"
 repos:
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.1.3
+    rev: v0.1.7
     hooks:
       - id: ruff
         args: [--fix]
@@ -23,18 +23,18 @@ repos:
         additional_dependencies: [black]
         exclude: ^(README.md|paper/paper.md)$
   - repo: https://github.com/pycqa/flake8
-    rev: 6.0.0
+    rev: 6.1.0
     hooks:
       - id: flake8
         entry: pflake8
         files: ^src/
         additional_dependencies:
-          - pyproject-flake8==6.0.0
-          - flake8-bugbear==22.12.6
-          - flake8-typing-imports==1.14.0
-          - flake8-docstrings==1.6.0
-          - flake8-rst-docstrings==0.3.0
-          - flake8-rst==0.8.0
+          - pyproject-flake8
+          - flake8-bugbear
+          - flake8-typing-imports
+          - flake8-docstrings
+          - flake8-rst-docstrings
+          - flake8-rst
   - repo: https://github.com/pre-commit/pygrep-hooks
     rev: v1.10.0
     hooks:
@@ -43,7 +43,7 @@ repos:
       - id: rst-directive-colons
       - id: rst-inline-touching-normal
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.6.1
+    rev: v1.7.1
     hooks:
       - id: mypy
         files: ^src/
