### 2023-12-08 3f607e1
add explicit stacklevel to warnings (fixes flake8 B028)
