### 2023-12-08 d3c690d
fix dead dynamic wf doc link reported in https://github.com/openjournals/joss-reviews/issues/5995#issuecomment-1847542926
