### 2023-12-08 0906788
ruff auto-remove needless lambdas
