### 2023-12-08 7666ca6
CI check for dead links in jupyter notebooks and markdown docs
