+++ pyproject.toml
@@ -8,7 +8,7 @@ description = "jobflow is a library for writing computational workflows"
 readme = "README.md"
 keywords = ["high-throughput", "workflow"]
 license = { text = "modified BSD" }
-authors = [{ name = "Alex Ganose", email = "alexganose@gmail.com" }]
+authors = [{ name = "Alex Ganose", email = "a.ganose@imperial.ac.uk" }]
 dynamic = ["version"]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
