+++ .github/workflows/deploy.yml
@@ -10,7 +10,7 @@ jobs:
   deploy-docs:
 
     # only run if commit is a push to master and the testing finished
-    if: ${{ github.event.workflow_run.conclusion == 'success' && github.event.workflow_run.event == 'push' && startsWith(github.event.workflow_run.head_branch, 'v0.') }}
+    if: ${{ github.repository_owner == 'materialsproject' && github.event.workflow_run.conclusion == 'success' && github.event.workflow_run.event == 'push' && startsWith(github.event.workflow_run.head_branch, 'v0.') }}
     runs-on: ubuntu-latest
 
     steps:
@@ -44,7 +44,7 @@ jobs:
   deploy-pypi:
 
     # only run if commit is tagged as a version and the docs finished
-    if: ${{ startsWith(github.event.workflow_run.head_branch, 'v') }}
+    if: github.repository_owner == 'materialsproject' && github.event.workflow_run.conclusion == 'success' && github.event.workflow_run.event == 'push' && startsWith(github.event.workflow_run.head_branch, 'v0.')
     runs-on: ubuntu-latest
     needs:
       - deploy-docs
