### 2023-12-11 #5995 @Andrew-S-Rosen
@rashatwi (CC @arfon):

Thank you very much for your very detailed review! We are glad to hear that you have found Jobflow both easy to use and potentially valuable in your own future research.

> The link in this part of the documentation (Defining jobs in jobflow) is not working. Perhaps you were trying to refer to this section instead:

This has now been addressed in [PR 503](https://github.com/materialsproject/jobflow/pull/503). We also added an automatic link-checker to ensure that such an issue doesn't happen again.

> Introduction to Jobflow tutorial: After executing flow = Flow( jobs=[time_github, time_google, time_nyt, sum_times], output=sum_times.output )...

Thank you for your report. This was an artifact of running the same cell again in a Jupyter Notebook. [PR 505](https://github.com/materialsproject/jobflow/pull/505) provide a bit of guidance to ensure that this doesn't happen to other users going through the tutorial.

> I found the tutorials provided very basic and some are repetitive, which might be intentional. However, I would appreciate the addition of a bit more advanced examples, e.g., there is not much about setting a custom store other than MemoryStore. The docstrings provided are a very good starting point, but I’m not sure if I’m missing any additional information out there about the custom stores.

Indeed, we strived to keep the Jobflow documentation quite accessible and domain-agnostic, with ample detail provided in the API.

The information about more "advanced" stores is admittedly scattered across a few repositories. We have streamlined things in the Jobflow documentation to help users like yourself. This has now been addressed in [PR 506](https://github.com/materialsproject/jobflow/pull/506).

> One of the main features of Jobflow is being independent of the workflow manager and you already have an adapter/support for FireWorks. I was wondering if you are planning to add additional adapters, and if so, which ones? If not and if this is a task left to the user to create their own adapters, it would be nice to have an example/guide on the website on how to get started.

This is a good question. Recently, several of the co-authors on this work have developed a custom workflow manager built around Jobflow called `jobflow-remote`, which we anticipate will be made public in the near future. Additionally, [PR 182](https://github.com/materialsproject/jobflow/pull/182) contains a work-in-progress adapter for the Myqueue workflow manager. The existing [FireWorks adapter](https://github.com/materialsproject/jobflow/blob/main/src/jobflow/managers/fireworks.py), in addition to the aforementioned example, should hopefully provide some guidance on what is needed to create a custom adapter. Since each workflow manager operates differently, it is difficult to provide a uniform set of guidelines, but starting from the FireWorks example should get the user most of the way there.

> Is there support in the Job output for a wider range of datatypes, not only limited to Monty serializable datatypes?

Due to the way in which Jobflow is written, all objects must be (de)serializable by Monty. We note that this includes [`MSONable`](https://github.com/materialsvirtuallab/monty/blob/5d37679e1f0cd58a8d165d76f3a095cb71266831/monty/json.py#L92-L131) objects as well as virtually all Python primitives (e.g. lists, dictionaries, strings, floats, etc.). In practice, any object can be made `MSONable`, such as by adding a custom `.as_dict()` or `.to_dict()` property to define a custom routine on how to (de)serialize the object.

> Having rewritten atomate2 based on atomate, how difficult was the transition from FireWorks to Jobflow for defining the flows, i.e., what was the overhead time for modifying/building upon existing workflows that utilize the FireWorks structure to the use Jobflow instead. I understand that this is a very subjective question, but any comment about that would be really appreciated. I would imagine there are lots of people who have been using the FireWorks package for defining workflows and might be interested in making this transition as well.

First, I'll answer the question independent of atomate2. In general, we anticipate that the transition from a FireWorks-based library to one centered around Jobflow should be relatively straightforward. The use of the simple `@job` decorator makes it very easy to turn a given compute function into a discrete unit of work, and this is much more natural for developers to think about than coupling the code to FireWorks directly. There are certainly some learning curves, e.g. to understand how to handle the `Response` (dynamic job) object, but this is expected.

For your specific question about atomate-->atomate2, however, the story is a bit different. That's because atomate1 is a very large, complex package and atomate2 was designed to address this by introducing a variety of new concepts that are built around Jobflow but are not specific to Jobflow itself. Porting workflows from atomate1 to atomate2 was not a simple translation process as a result --- it largely is a conceptual rewrite. For this reason, the initial time commitment has been high, although we stress that we believe the long-term payoff is high as well since the workflows are much more streamlined, more maintainable, and easier to test. If you're simply interested in taking an existing package and swapping out the workflow bits to be Jobflow-centric, that process is significantly more straightforward to do.
