+++ pyproject.toml
@@ -61,7 +61,7 @@ strict = [
     "pydantic==2.5.2",
     "pydash==7.0.6",
     "pydot==1.4.2",
-    "typing-extensions==4.8.0",
+    "typing-extensions==4.9.0",
 ]
 
 [project.urls]
