### 2023-12-18 534526c
Bump maggma from 0.59.0 to 0.60.0
Bumps [maggma](https://github.com/materialsproject/maggma) from 0.59.0 to 0.60.0.
- [Release notes](https://github.com/materialsproject/maggma/releases)
- [Changelog](https://github.com/materialsproject/maggma/blob/main/docs/CHANGELOG.md)
- [Commits](https://github.com/materialsproject/maggma/compare/v0.59.0...v0.60.0)

---
updated-dependencies:
- dependency-name: maggma
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
