### 2023-12-18 ea2b831
Bump moto from 4.2.11 to 4.2.12
Bumps [moto](https://github.com/getmoto/moto) from 4.2.11 to 4.2.12.
- [Release notes](https://github.com/getmoto/moto/releases)
- [Changelog](https://github.com/getmoto/moto/blob/master/CHANGELOG.md)
- [Commits](https://github.com/getmoto/moto/compare/4.2.11...4.2.12)

---
updated-dependencies:
- dependency-name: moto
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
