### 2023-12-25 4a4992f
Bump ipython from 8.18.1 to 8.19.0
Bumps [ipython](https://github.com/ipython/ipython) from 8.18.1 to 8.19.0.
- [Release notes](https://github.com/ipython/ipython/releases)
- [Commits](https://github.com/ipython/ipython/compare/8.18.1...8.19.0)

---
updated-dependencies:
- dependency-name: ipython
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
