### 2024-01-01 63263f1
Bump pydot from 1.4.2 to 2.0.0
Bumps [pydot](https://github.com/pydot/pydot) from 1.4.2 to 2.0.0.
- [Changelog](https://github.com/pydot/pydot/blob/master/ChangeLog)
- [Commits](https://github.com/pydot/pydot/compare/v1.4.2...v2.0.0)

---
updated-dependencies:
- dependency-name: pydot
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
