+++ .github/workflows/deploy.yml
@@ -6,49 +6,10 @@ on:
     types: [completed]
 
 jobs:
-
-  deploy-docs:
-
-    # only run if commit is a push to master and the testing finished
-    if: ${{ github.repository_owner == 'materialsproject' && github.event.workflow_run.conclusion == 'success' && github.event.workflow_run.event == 'push' && startsWith(github.event.workflow_run.head_branch, 'v0.') }}
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          ref: ${{ github.event.workflow_run.head_branch }}
-
-      - name: Install pandoc
-        run: sudo apt-get install pandoc
-
-      - uses: actions/setup-python@v4
-        with:
-          python-version: '3.10'
-          cache: pip
-          cache-dependency-path: pyproject.toml
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install .[strict,docs]
-
-      - name: Build
-        run: sphinx-build docs docs_build
-
-      - name: Deploy
-        uses: peaceiris/actions-gh-pages@v3
-        with:
-          deploy_key: ${{ secrets.ACTIONS_DEPLOY_KEY }}
-          publish_dir: ./docs_build
-
   deploy-pypi:
-
-    # only run if commit is tagged as a version and the docs finished
+    # only run if commit is a push to master, the testing finished, and tagged as version
     if: github.repository_owner == 'materialsproject' && github.event.workflow_run.conclusion == 'success' && github.event.workflow_run.event == 'push' && startsWith(github.event.workflow_run.head_branch, 'v0.')
     runs-on: ubuntu-latest
-    needs:
-      - deploy-docs
-
     steps:
       - uses: actions/checkout@v3
         with:
@@ -70,11 +31,9 @@ jobs:
           password: ${{ secrets.PYPI_PASSWORD }}
 
   deploy-github:
-
     runs-on: ubuntu-latest
     needs:
       - deploy-pypi
-
     steps:
       - uses: actions/checkout@v3
         with:
