### 2024-01-07 #5995 @utf
Thanks @arfon, I've released a new version of jobflow (v0.1.16) on [PyPI](https://pypi.org/project/jobflow/0.1.16/) and [GitHub](https://github.com/materialsproject/jobflow/releases/tag/v0.1.16) and archived it on [Zenodo](https://zenodo.org/records/10466868). The DOI for zenodo is [10.5281/zenodo.10466868](https://doi.org/10.5281/zenodo.10466868).
