### 2024-01-07 #5995 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
